+++ src/d3/quadFeature.js
@@ -196,8 +196,8 @@ var d3_quadFeature = function (arg) {
    */
   this._update = function () {
     s_update.call(m_this);
-    if (m_this.buildTime().getMTime() <= m_this.dataTime().getMTime() ||
-        m_this.buildTime().getMTime() < m_this.getMTime()) {
+    if (m_this.buildTime().timestamp() <= m_this.dataTime().timestamp() ||
+        m_this.buildTime().timestamp() < m_this.timestamp()) {
       m_this._build();
     }
     return m_this;