+++ src/gl/choroplethFeature.js
@@ -71,8 +71,8 @@ var gl_choroplethFeature = function (arg) {
    */
   this._update = function () {
     s_update.call(m_this);
-    if (m_this.dataTime().getMTime() >= m_this.buildTime().getMTime() ||
-        m_this.updateTime().getMTime() <= m_this.getMTime()) {
+    if (m_this.dataTime().timestamp() >= m_this.buildTime().timestamp() ||
+        m_this.updateTime().timestamp() <= m_this.timestamp()) {
       m_this._wipePolygons();
       m_this._build();
     }