+++ src/isolineFeature.js
@@ -536,7 +536,7 @@ var isolineFeature = function (arg) {
    * @returns {this}
    */
   this.labelPositions = function () {
-    if (m_this.dataTime().getMTime() >= m_this.buildTime().getMTime()) {
+    if (m_this.dataTime().timestamp() >= m_this.buildTime().timestamp()) {
       m_this._build();
     }
     m_lastLabelPositions = null;
@@ -756,8 +756,8 @@ var isolineFeature = function (arg) {
   this._update = function () {
     s_update.call(m_this);
 
-    if (m_this.dataTime().getMTime() >= m_this.buildTime().getMTime() ||
-        m_this.updateTime().getMTime() <= m_this.getMTime()) {
+    if (m_this.dataTime().timestamp() >= m_this.buildTime().timestamp() ||
+        m_this.updateTime().timestamp() <= m_this.timestamp()) {
       m_this._build();
     }
     m_this.updateTime().modified();