+++ tests/cases/mapInteractor.js
@@ -1352,7 +1352,7 @@ describe('mapInteractor', function () {
     expect(map.info.panArgs.y).toBe(0);
     interactor.simulateEvent(
       'mouseup', {map: {x: 10, y: 0}, button: 'left'});
-    start = new Date().getTime();
+    start = Date.now();
     stepAnimationFrame(start);
     expect(map.info.pan).toBe(2);
     expect(map.info.panArgs.x).toBeGreaterThan(0.25);
@@ -1412,7 +1412,7 @@ describe('mapInteractor', function () {
     expect(map.info.panArgs.x).toBeCloseTo(1);
     interactor.simulateEvent(
       'mouseup', {map: {x: 200.1, y: 0}, button: 'left'});
-    start = new Date().getTime();
+    start = Date.now();
     stepAnimationFrame(start);
     expect(map.info.pan).toBe(4);
     expect(map.info.panArgs.x).toBeLessThan(0);
@@ -1449,7 +1449,7 @@ describe('mapInteractor', function () {
         'wheel',
         {wheelDelta: {x: 0, y: -20}, wheelMode: 0}
       );
-      start = new Date().getTime();
+      start = Date.now();
       stepAnimationFrame(start);
       expect(map.zoom()).toBe(2);
       stepAnimationFrame(start + 50);
@@ -1469,7 +1469,7 @@ describe('mapInteractor', function () {
         'wheel',
         {wheelDelta: {x: 0, y: -20}, wheelMode: 0}
       );
-      start = new Date().getTime();
+      start = Date.now();
       stepAnimationFrame(start);
       expect(map.zoom()).toBe(2);
       stepAnimationFrame(start + 50);
@@ -1501,7 +1501,7 @@ describe('mapInteractor', function () {
         'wheel',
         {wheelDelta: {x: 0, y: -20}, wheelMode: 0}
       );
-      start = new Date().getTime();
+      start = Date.now();
       stepAnimationFrame(start);
       expect(map.zoom()).toBe(2);
       stepAnimationFrame(start + 50);
@@ -1535,7 +1535,7 @@ describe('mapInteractor', function () {
         'wheel',
         {wheelDelta: {x: 0, y: -20}, wheelMode: 0}
       );
-      start = new Date().getTime();
+      start = Date.now();
       stepAnimationFrame(start);
       expect(map.zoom()).toBe(2);
       stepAnimationFrame(start + 50);